requirements.txt
appdirs==1.4.3
Autologging==1.1.0
backports.csv==1.0.4
beautifulsoup4==4.5.3
configparser==3.5.0
coverage==4.3.4
coveralls==1.1
Django==1.11
django-braces==1.11.0
django-react-templatetags==300
django-rest-framework==0.1.0
django-search-views==0.3.7
django-select2-forms==2.0.2
django-sortedm2m==1.5.0
djangorestframework==3.7.3
docopt==0.6.2
enum34==1.1.6
ez-setup==0.9
flake8==3.3.0
flake8-docstrings==1.0.3
flake8-polyfill==1.0.1
funcsigs==1.0.2
html5lib==0.999999999
mccabe==0.6.1
mock==2.0.0
packaging==16.8
pbr==2.0.0
psycopg2==2.7.1
pycodestyle==2.3.1
pydocstyle==1.1.1
pyflakes==1.5.0
pyparsing==2.2.0
python-coveralls==2.9.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
requests-mock==1.3.0
six==1.10.0
tqdm==4.14.0
webencodings==0.5.1
